[#7463] Fix Reactor-Netty HTTP client plugin
